+++ package.json
@@ -10,9 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^1.9.0",
     "@tanstack/react-query": "^5.80.6",
     "@uiw/react-md-editor": "^4.0.7",
+    "axios": "^1.9.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
